requirements.txt
absl-py==2.3.0
accelerate==1.7.0
accelerated-scan==0.2.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alabaster==1.0.0
alembic==1.16.1
aniso8601==10.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asciitree==0.3.3
asttokens==3.0.0
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.1
bitsandbytes==0.45.3
black==24.10.0
blinker==1.9.0
boto3==1.38.24
botocore==1.38.24
braceexpand==0.1.7
Brotli==1.1.0
cdifflib==1.2.6
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
clip==0.2.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
coverage==7.8.2
cryptography==45.0.3
cycler==0.12.1
cytoolz==1.0.1
dataclasses-json==0.6.7
DataProperty==1.1.0
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.33.1
dill==0.3.8
Distance==0.1.3
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
editdistance==0.8.1
einops==0.8.1
einops-exts==0.0.4
emoji==2.14.1
evaluate==0.4.3
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fasteners==0.19
fiddle==0.3.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
Flask-RESTful==0.3.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
g2p-en==2.1.0
gdown==5.2.0
gevent==25.5.1
geventhttpclient==2.0.2
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.2.2
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.27.0
huggingface-hub==0.32.2
hydra-core==1.3.2
idna==3.10
ijson==3.4.0
imageio==2.37.0
imagesize==1.4.1
immutabledict==4.2.0
importlib_metadata==8.6.1
inflect==7.5.0
iniconfig==2.1.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.36.0
isort==5.13.2
itsdangerous==2.2.0
Janome==0.5.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jiwer==3.1.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
langdetect==1.0.9
latexcodec==3.0.0
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.30.3
libcst==1.8.0
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdurl==0.1.2
mediapy==1.1.6
megatron-core==0.12.1
megatron-energon==5.2.0
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multi-storage-client==0.20.3
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nemo-toolkit==2.3.1
nemo_text_processing==1.1.0
nerfacc==0.5.3
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numcodecs==0.13.1
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-modelopt==0.27.1
nvidia-modelopt-core==0.27.1
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvidia-pytriton==0.5.14
nvidia-resiliency-ext==0.4.0
nvidia_lm_eval==25.4.1
omegaconf==2.3.0
onnx==1.17.0
open-clip-torch==2.24.0
openai==1.61.0
OpenCC==1.1.9
opentelemetry-api==1.33.1
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.3
peft==0.15.2
pesq==0.0.4
pexpect==4.9.0
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
portalocker==3.1.1
prettytable==3.16.0
progress==1.6
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
PuLP==3.1.1
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==20.0.0
pybind11==2.13.6
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
pynini==2.1.6.post1
pynvml==12.0.0
pyparsing==3.2.3
pypdf==5.5.0
pypinyin==0.54.0
pypinyin-dict==0.9.0
PySocks==1.7.1
pystoi==0.4.1
pytablewriter==1.2.1
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.1
pytest-random-order==1.1.1
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-rapidjson==1.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.11
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.25.1
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
s3fs==0.4.2
s3transfer==0.13.0
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
sh==2.2.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
sox==1.5.0
soxr==0.5.0.post1
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.3
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
taming-transformers==0.0.1
tcolorpy==0.1.7
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorstore==0.1.71
termcolor==3.1.0
text-unidecode==1.3
textdistance==4.6.3
texterrors==0.5.1
threadpoolctl==3.6.0
tiktoken==0.7.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
toolz==1.0.0
torch==2.7.0
torchaudio==2.7.0
torchdiffeq==0.2.5
torchmetrics==1.7.1
torchprofile==0.0.4
torchsde==0.2.6
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
trampoline==0.1.2
transformers==4.51.3
trimesh==4.6.10
triton==3.3.0
tritonclient==2.51.0
typeguard==4.4.2
typepy==1.3.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.14.9
unstructured-client==0.35.0
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.19.11
wcwidth==0.2.13
webdataset==0.2.111
Werkzeug==3.1.3
wget==3.2
word2number==1.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zarr==2.18.3
zipp==3.22.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
